chore: clippy clean (0 warnings) + revert wire-compat detour
Workspace clippy lint pass:
- Box VerifyOutcome::Ok variant + extract VerifyOk struct (large_enum_variant)
- Box Command::Serve(ServeFlags) variant (large_enum_variant)
- AppState init via struct-update syntax (field_reassign_with_default)
- ExpiredPendingConfirm type alias (type_complexity in db.rs)
- &PathBuf → &Path on closure_hash_from_path (ptr_arg)
- let-else → ? in dispatch_target_for_checkin (question_mark)
- doc_lazy_continuation: 14 fixes across reconciler/lib.rs,
  control-plane/db.rs/issuance.rs/server/mod.rs, release/lib.rs
- Auto-fixes: Ok-?-trip removal in agent/comms.rs, needless_borrows
  in 5 integration tests, unnecessary_cast in agent/main.rs,
  match-like-matches in handlers.rs, unnecessary parens

Reverts the wire-compat detour from 8236d24:
- Drop synthesize-strict from mk-fleet (clean schema: mode-only)
- Drop strict from proto Compliance (clean schema)
- Lab will be redeployed manually rather than via wire-compat dance
- mk-fleet goldens revert to clean strict-less form

Workspace: 245 tests passing, 0 clippy warnings.

+++ crates/nixfleet-agent/src/main.rs
@@ -254,7 +254,7 @@ async fn main() -> anyhow::Result<()> {
         // ticker's regular cadence). The ticker fires on its own
         // schedule; backoff is layered.
         if consecutive_failures > 0 {
-            let multiplier = (1u64 << (consecutive_failures.min(3))) as u64; // 1, 2, 4, 8
+            let multiplier = 1u64 << (consecutive_failures.min(3)); // 1, 2, 4, 8
             let base = args.poll_interval.saturating_mul(multiplier);
             // ±20% jitter.
             let jitter_pct: f64 = {
